--HG--
branch : soc.2013.gobjectification.plugins

+++ libpurple/protocols/jabber/jabber.h
@@ -316,7 +316,7 @@ typedef struct _JabberIdentity
 typedef struct _JabberBytestreamsStreamhost {
 	char *jid;
 	char *host;
-	int port;
+	guint16 port;
 	char *zeroconf;
 } JabberBytestreamsStreamhost;
